+++ app/src/main/java/com/kazu/garbageday/common/utils/ColorUtil.kt
@@ -1,0 +1,4 @@
+package com.kazu.garbageday.common.utils
+
+class ColorUtil {
+}
