+++ app/src/main/java/com/kazu/garbageday/common/utils/ToolbarUtil.kt
@@ -1,0 +1,4 @@
+package com.kazu.garbageday.common.utils
+
+object ToolbarUtil {
+}
