+++ app/src/main/java/com/kazu/garbageday/ui/garbageDetail/GarbageDetailFragment.kt
@@ -1,0 +1,4 @@
+package com.kazu.garbageday.ui.GarbageDatail
+
+class GarbageDatailFragment {
+}
