+++ app/src/main/java/com/kazu/garbageday/ui/garbageDetail/GarbageDetailViewModel.kt
@@ -1,0 +1,4 @@
+package com.kazu.garbageday.ui.garbageDetail
+
+class GarbageDetailViewModel {
+}
