+++ app/src/main/java/com/kazu/garbageday/ui/garbageList/adapter/GarbageMajorItemsAdapter.kt
@@ -1,0 +1,4 @@
+package com.kazu.garbageday.ui.garbageList.adapter
+
+class GarbageMajorItemsAdapter {
+}
