+++ app/src/main/res/layout/burning_garbage_tag.xml
@@ -1,0 +1,68 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:orientation="vertical">
+
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="月曜日"
+        android:textAlignment="center"
+        android:textStyle="bold"
+        android:textSize="18sp" />
+</LinearLayout>
+
+<LinearLayout
+android:layout_width="match_parent"
+android:layout_height="match_parent"
+android:orientation="horizontal">
+
+<LinearLayout
+    android:id="@+id/text_home2"
+    android:layout_width="wrap_content"
+    android:layout_height="50dp"
+    android:layout_marginStart="4dp"
+    android:background="@drawable/shape_rounded_corners"
+    android:orientation="vertical">
+
+    <TextView
+        android:id="@+id/ad2"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="8dp"
+        android:layout_marginEnd="8dp"
+        android:text="燃えるゴミ"
+        android:textAlignment="center"
+        android:textSize="20sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/text_home" />
+</LinearLayout>
+
+<LinearLayout
+    android:id="@+id/text_home"
+    android:layout_width="wrap_content"
+    android:layout_height="50dp"
+    android:layout_marginStart="4dp"
+    android:background="@drawable/shape_rounded_corners"
+    android:orientation="vertical">
+
+    <TextView
+        android:id="@+id/ad"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="8dp"
+        android:layout_marginEnd="8dp"
+        android:text="粗大ゴミ"
+        android:textAlignment="center"
+        android:textSize="20sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/text_home" />
+</LinearLayout>
+
+</LinearLayout>
