+++ app/src/main/res/layout/no_burning_garbage_tag.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/text_home2"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_marginStart="4dp"
+    android:layout_marginBottom="4dp"
+    android:background="@drawable/burning_garbage_background"
+    android:orientation="vertical">
+
+    <TextView
+        android:id="@+id/ad2"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginEnd="8dp"
+        android:paddingTop="6dp"
+        android:paddingBottom="6dp"
+        android:text="燃えるゴミ"
+        android:textColor="@color/color_ffffff"
+        android:textSize="15dp" />
+</LinearLayout>
