+++ app/src/main/res/layout/small_metal_tag.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/text_home2"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_marginStart="4dp"
+    android:background="@drawable/plastic_trash_background"
+    android:orientation="vertical">
+
+    <TextView
+        android:id="@+id/ad2"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginEnd="8dp"
+        android:paddingTop="6dp"
+        android:paddingBottom="6dp"
+        android:text="プラスチック資源"
+        android:textColor="@color/color_0000cc"
+        android:textSize="15dp" />
+</LinearLayout>
